Update failure returns to 0 from -1 for int to unsigned int changes

+++ pkg/utils/converters.go
@@ -29,39 +29,39 @@ type HashrateUnit int
 
 // ConvertHashrate takes an input hashrate, as well as an input and output HashrateUnit, and converts the input to the
 // output hashrate with the given unit conversion. Returns the converted hashrate as an integer.
-func ConvertHashrate(inputHashrate int, inputHashrateUnit HashrateUnit, outputHashrateUnit HashrateUnit) int {
+func ConvertHashrate(inputHashrate uint, inputHashrateUnit HashrateUnit, outputHashrateUnit HashrateUnit) uint {
 	// First we convert the input to hashes/second
-	hashesPerSecond := 0
+	hashesPerSecond := uint(0)
 
 	switch inputHashrateUnit {
 	case HashesPerSecond:
-		hashesPerSecond = inputHashrate * int(math.Pow10(BasePow10Exponential))
+		hashesPerSecond = inputHashrate * uint(math.Pow10(BasePow10Exponential))
 	case KiloHashesPerSecond:
-		hashesPerSecond = inputHashrate * int(math.Pow10(KiloPow10Exponential))
+		hashesPerSecond = inputHashrate * uint(math.Pow10(KiloPow10Exponential))
 	case MegaHashesPerSecond:
-		hashesPerSecond = inputHashrate * int(math.Pow10(MegaPow10Exponential))
+		hashesPerSecond = inputHashrate * uint(math.Pow10(MegaPow10Exponential))
 	case GigaHashesPerSecond:
-		hashesPerSecond = inputHashrate * int(math.Pow10(GigaPow10Exponential))
+		hashesPerSecond = inputHashrate * uint(math.Pow10(GigaPow10Exponential))
 	case TeraHashesPerSecond:
-		hashesPerSecond = inputHashrate * int(math.Pow10(TeraPow10Exponential))
+		hashesPerSecond = inputHashrate * uint(math.Pow10(TeraPow10Exponential))
 	case PetaHashesPerSecond:
-		hashesPerSecond = inputHashrate * int(math.Pow10(PetaPow10Exponential))
+		hashesPerSecond = inputHashrate * uint(math.Pow10(PetaPow10Exponential))
 	}
 
 	// Now convert to output units
 	switch outputHashrateUnit {
 	case HashesPerSecond:
-		return hashesPerSecond / int(math.Pow10(BasePow10Exponential))
+		return hashesPerSecond / uint(math.Pow10(BasePow10Exponential))
 	case KiloHashesPerSecond:
-		return hashesPerSecond / int(math.Pow10(KiloPow10Exponential))
+		return hashesPerSecond / uint(math.Pow10(KiloPow10Exponential))
 	case MegaHashesPerSecond:
-		return hashesPerSecond / int(math.Pow10(MegaPow10Exponential))
+		return hashesPerSecond / uint(math.Pow10(MegaPow10Exponential))
 	case GigaHashesPerSecond:
-		return hashesPerSecond / int(math.Pow10(GigaPow10Exponential))
+		return hashesPerSecond / uint(math.Pow10(GigaPow10Exponential))
 	case TeraHashesPerSecond:
-		return hashesPerSecond / int(math.Pow10(TeraPow10Exponential))
+		return hashesPerSecond / uint(math.Pow10(TeraPow10Exponential))
 	case PetaHashesPerSecond:
-		return hashesPerSecond / int(math.Pow10(PetaPow10Exponential))
+		return hashesPerSecond / uint(math.Pow10(PetaPow10Exponential))
 	}
 
 	return -1
